# app/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_aeciopires_mytoolkit_internal_tools_k8svalidate.DocumentResult:
    properties:
      api_version:
        type: string
      error:
        type: string
      index:
        type: integer
      kind:
        type: string
      name:
        type: string
      valid:
        type: boolean
    type: object
  github_com_aeciopires_mytoolkit_internal_tools_k8svalidate.Result:
    properties:
      documents:
        items:
          $ref: '#/definitions/github_com_aeciopires_mytoolkit_internal_tools_k8svalidate.DocumentResult'
        type: array
      valid:
        type: boolean
    type: object
...
  internal_cli.ToolErrorBody:
    properties:
      code:
        example: INVALID_OPTIONS
        type: string
      message:
        example: invalid options object
        type: string
    type: object
  internal_cli.ToolErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_cli.ToolErrorBody'
      success:
        example: false
        type: boolean
    type: object
  internal_cli.ToolMeta:
    properties:
      duration_ms:
        example: 0.08
        type: number
      tool:
        example: base64
        type: string
    type: object
...
info:
  contact:
    name: Aecio Pires
    url: https://github.com/aeciopires
  description: Developer utilities (JSON/YAML formatting and conversion, JWT, hashing,
    encoding, password generation, QR codes, and more) exposed as a REST API. Every
    tool also has a CLI subcommand and a web UI backed by this same API — see the
    project README for the full picture.
  license:
    name: MIT
    url: https://github.com/aeciopires/mytoolkit/blob/main/LICENSE
  title: MyToolkit REST API
  version: "1.0"
paths:
...
  /api/v1/tools/k8s-validate:
    post:
      consumes:
      - application/json
      description: 'Validates a YAML document (or a "---"-separated multi-document
        stream) against the two fields every Kubernetes API object requires: non-empty
        apiVersion and kind, plus an object-shaped metadata if present. Does not validate
        against any specific resource''s full schema. HTTP 200 with "data.valid":false
        means the manifests parsed but aren''t valid; only a hard YAML syntax error
        returns a non-2xx status.'
      parameters:
      - description: YAML document or multi-document stream
        in: body
        name: request
        required: true
        schema:
          properties:
            input:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/github_com_aeciopires_mytoolkit_internal_tools_k8svalidate.Result'
              meta:
                $ref: '#/definitions/internal_cli.ToolMeta'
              success:
                type: boolean
            type: object
        "400":
          description: e.g. INVALID_YAML (syntax error), NO_DOCUMENTS
          schema:
            $ref: '#/definitions/internal_cli.ToolErrorResponse'
      summary: Validate Kubernetes-shaped YAML
      tags:
      - tools
...
schemes:
- http
- https
swagger: "2.0"